Entity System Event Emitting (#67)
- Shift the way systems can react to components being added and removed from entities. Now, `ecs::Registry` can be given a callback with signature `void(ev::Event&)` which is called whenever a component is added or removed. When either of these happens, the callback is called with an event containing either a `AddedEvent<Component>` or a `RemovedEvent<Component>` where `Component` is a template parameter. The `Scene` class passes a callback in its constructor so that these events get sent to systems. So instead of systems registering callbacks with the `ecs::Registry`, they can simply react to events in their `OnEvent` functions.
- Removed `OnAdd` and `OnRemove` from the `ecs::Registry` as they are no longer needed. Removed the vector of callbacks associated with them.
- Inside the component map in `ecs::Registry`, there is now a `std::function<ev::Event(ecs::Entity)>` which can be called to generate the `RemovedEvent<Component>`. This is necessary since the type of the component is completely erased at removal time, so we construct this function at addition time which can retain the type.
- Added `ecs::Regisry::emit(ev::Event&)`, which allows users to invoke the callback with custom events. The intended purpose here is to allow systems to send out events to other systems.
- `PhysicsEngine3D` now emits collision events containing the two entities involved in the collision.
- `ScriptRunner` has been modified to allow scripts to implement a `OnCollisionEvent` function. Sadly there is no scalable way to let scripts access custom events; they must be manually implemented in the `ScriptRunner`. Hopefully a better solution is found later.
- A toy example has been added to `Anvil`. There is now a bouncing ball in the scene with a `LightComponent`. Every time it makes contact with another entity, the colour of the light changes.
- Systems now receive the `ecs::Registry` by reference instead of the `Scene`. Only the registry was accessed by current systems; and the only reason the whole scene was given was because the original idea was that systems would communicate via direct access from the scene. This was a poor design however, and the new event driven approach is much better.
- With an increasing number of event types, the number of events emitted per frame is going to increase, the next job will be to allow systems to subscribe to events so that they only receive what they want, although this will probably require another redesign. Nevertheless, the current model is much clear than the previous one.
- Added `AreEntitiesEqual` function to Lua

+++ Sprocket/Scene/Systems/PhysicsEngine3D.cpp
@@ -1,4 +1,5 @@
 #include "PhysicsEngine3D.h"
+#include "ECS.h"
 #include "Log.h"
 #include "Scene.h"
 #include "Components.h"
@@ -84,6 +85,37 @@ public:
     float Fraction() const { return d_fraction; }
 };
 
+class EventListener : public rp3d::EventListener
+{
+    ecs::Registry* registry;
+
+public:
+    EventListener(ecs::Registry* r) : registry(r) {
+        assert(r);
+    }
+
+    void onContact(const rp3d::CollisionCallback::CallbackData& data) override
+    {
+        const auto name = [](const ecs::Entity& e) {
+            if (e.Has<NameComponent>()) {
+                return e.Get<NameComponent>().name;
+            }
+            return guid::Stringify(e.Id());
+        };
+
+        for (u32 p = 0; p != data.getNbContactPairs(); ++p) {
+            auto pair = data.getContactPair(p);
+            auto type = pair.getEventType();
+            if (type == rp3d::CollisionCallback::ContactPair::EventType::ContactStart) {
+                ecs::Entity e1 = *static_cast<ecs::Entity*>(pair.getBody1()->getUserData());
+                ecs::Entity e2 = *static_cast<ecs::Entity*>(pair.getBody2()->getUserData());
+                ev::Event event = ev::make_event<CollisionEvent>(e1, e2);
+                registry->emit(event);
+            }
+        }
+    }
+};
+
 }
 
 struct EntityData
@@ -101,6 +133,8 @@ struct PhysicsEngine3DImpl
 {
     rp3d::PhysicsCommon pc;
     rp3d::PhysicsWorld* world;
+
+    std::unique_ptr<EventListener> listener;
 
     float lastFrameLength = 0;
 
@@ -150,36 +184,36 @@ PhysicsEngine3D::PhysicsEngine3D(const glm::vec3& gravity)
 {
 }
 
-void PhysicsEngine3D::OnStartup(Scene& scene)
+void PhysicsEngine3D::OnEvent(ecs::Registry& registry, ev::Event& event)
 {
-    scene.Entities().OnAdd<RigidBody3DComponent>([&](ecs::Entity entity) {
-        assert(entity.Has<Transform3DComponent>());
-        auto& tc = entity.Get<Transform3DComponent>();
-        auto& rc = entity.Get<RigidBody3DComponent>();
+    if (auto data = event.get_if<ecs::ComponentAddedEvent<RigidBody3DComponent>>()) {
+        assert(data->entity.Has<Transform3DComponent>());
+        auto& tc = data->entity.Get<Transform3DComponent>();
+        auto& rc = data->entity.Get<RigidBody3DComponent>();
 
-        auto& entry = d_impl->entityData[entity];
-        entry.entity = entity;
+        auto& entry = d_impl->entityData[data->entity];
+        entry.entity = data->entity;
         entry.body = d_impl->world->createRigidBody(Convert(tc));
         entry.body->setUserData(static_cast<void*>(&entry.entity));
-    });
+    }
 
-    scene.Entities().OnRemove<RigidBody3DComponent>([&](ecs::Entity entity) {
-        entity.Remove<BoxCollider3DComponent>();
-        entity.Remove<SphereCollider3DComponent>();
-        entity.Remove<CapsuleCollider3DComponent>();
+    else if (auto data = event.get_if<ecs::ComponentRemovedEvent<RigidBody3DComponent>>()) {
+        data->entity.Remove<BoxCollider3DComponent>();
+        data->entity.Remove<SphereCollider3DComponent>();
+        data->entity.Remove<CapsuleCollider3DComponent>();
 
-        auto rigidBodyIt = d_impl->entityData.find(entity);
+        auto rigidBodyIt = d_impl->entityData.find(data->entity);
         d_impl->world->destroyRigidBody(rigidBodyIt->second.body);
         d_impl->entityData.erase(rigidBodyIt);
-    });
+    }
 
-    scene.Entities().OnAdd<BoxCollider3DComponent>([&](ecs::Entity entity) {
-        assert(entity.Has<Transform3DComponent>());
-        assert(entity.Has<RigidBody3DComponent>());
+    else if (auto data = event.get_if<ecs::ComponentAddedEvent<BoxCollider3DComponent>>()) {
+        assert(data->entity.Has<Transform3DComponent>());
+        assert(data->entity.Has<RigidBody3DComponent>());
 
-        auto& tc = entity.Get<Transform3DComponent>();
-        auto& bc = entity.Get<BoxCollider3DComponent>();
-        auto& entry = d_impl->entityData[entity];
+        auto& tc = data->entity.Get<Transform3DComponent>();
+        auto& bc = data->entity.Get<BoxCollider3DComponent>();
+        auto& entry = d_impl->entityData[data->entity];
 
         glm::vec3 dimensions = bc.halfExtents;
         if (bc.applyScale) { dimensions *= tc.scale; }
@@ -187,61 +221,67 @@ void PhysicsEngine3D::OnStartup(Scene& scene)
         rp3d::Transform transform = Convert(bc.position, bc.orientation);
 
         entry.boxCollider = entry.body->addCollider(shape, transform);
-        SetMaterial(entry.boxCollider, entity.Get<RigidBody3DComponent>());
-    });
+        SetMaterial(entry.boxCollider, data->entity.Get<RigidBody3DComponent>());
+    }
 
-    scene.Entities().OnRemove<BoxCollider3DComponent>([&](ecs::Entity entity) {
-        auto& entry = d_impl->entityData[entity];
+    else if (auto data = event.get_if<ecs::ComponentRemovedEvent<BoxCollider3DComponent>>()) {
+        auto& entry = d_impl->entityData[data->entity];
         entry.body->removeCollider(entry.boxCollider);
-    });
+    }
 
-    scene.Entities().OnAdd<SphereCollider3DComponent>([&](ecs::Entity entity) {
-        assert(entity.Has<Transform3DComponent>());
-        assert(entity.Has<RigidBody3DComponent>());
+    else if (auto data = event.get_if<ecs::ComponentAddedEvent<SphereCollider3DComponent>>()) {
+        assert(data->entity.Has<Transform3DComponent>());
+        assert(data->entity.Has<RigidBody3DComponent>());
 
-        auto& tc = entity.Get<Transform3DComponent>();
-        auto& sc = entity.Get<SphereCollider3DComponent>();
-        auto& entry = d_impl->entityData[entity];
+        auto& tc = data->entity.Get<Transform3DComponent>();
+        auto& sc = data->entity.Get<SphereCollider3DComponent>();
+        auto& entry = d_impl->entityData[data->entity];
         
         rp3d::SphereShape* shape = d_impl->pc.createSphereShape(sc.radius);
         rp3d::Transform transform = Convert(sc.position, sc.orientation);
 
         entry.sphereCollider = entry.body->addCollider(shape, transform);
-        SetMaterial(entry.sphereCollider, entity.Get<RigidBody3DComponent>());   
-    });
+        SetMaterial(entry.sphereCollider, data->entity.Get<RigidBody3DComponent>());   
+    }
 
-    scene.Entities().OnRemove<SphereCollider3DComponent>([&](ecs::Entity entity) {
-        auto& entry = d_impl->entityData[entity];
+    else if (auto data = event.get_if<ecs::ComponentRemovedEvent<SphereCollider3DComponent>>()) {
+        auto& entry = d_impl->entityData[data->entity];
         entry.body->removeCollider(entry.sphereCollider);
-    });
+    }
 
-    scene.Entities().OnAdd<CapsuleCollider3DComponent>([&](ecs::Entity entity) {
-        assert(entity.Has<Transform3DComponent>());
-        assert(entity.Has<RigidBody3DComponent>());
+    else if (auto data = event.get_if<ecs::ComponentAddedEvent<CapsuleCollider3DComponent>>()) {
+        assert(data->entity.Has<Transform3DComponent>());
+        assert(data->entity.Has<RigidBody3DComponent>());
 
-        auto& tc = entity.Get<Transform3DComponent>();
-        auto& cc = entity.Get<CapsuleCollider3DComponent>();
-        auto& entry = d_impl->entityData[entity];
+        auto& tc = data->entity.Get<Transform3DComponent>();
+        auto& cc = data->entity.Get<CapsuleCollider3DComponent>();
+        auto& entry = d_impl->entityData[data->entity];
         
         rp3d::CapsuleShape* shape = d_impl->pc.createCapsuleShape(cc.radius, cc.height);
         rp3d::Transform transform = Convert(cc.position, cc.orientation);
 
         entry.capsuleCollider = entry.body->addCollider(shape, transform);
-        SetMaterial(entry.capsuleCollider, entity.Get<RigidBody3DComponent>()); 
-    });
+        SetMaterial(entry.capsuleCollider, data->entity.Get<RigidBody3DComponent>()); 
+    }
 
-    scene.Entities().OnRemove<CapsuleCollider3DComponent>([&](ecs::Entity entity) {
-        auto& entry = d_impl->entityData[entity];
+    else if (auto data = event.get_if<ecs::ComponentRemovedEvent<CapsuleCollider3DComponent>>()) {
+        auto& entry = d_impl->entityData[data->entity];
         entry.body->removeCollider(entry.capsuleCollider);
-    });
+    }
 }
 
-void PhysicsEngine3D::OnUpdate(Scene& scene, double dt)
+void PhysicsEngine3D::OnStartup(ecs::Registry& registry)
+{
+    d_impl->listener = std::make_unique<EventListener>(&registry);
+    d_impl->world->setEventListener(d_impl->listener.get());
+}
+
+void PhysicsEngine3D::OnUpdate(ecs::Registry& registry, double dt)
 {
     // Pre Update
     // Do this even if not running so that the physics engine stays up
     // to date with the scene.
-    for (auto entity : scene.Entities().View<RigidBody3DComponent>()) {
+    for (auto entity : registry.View<RigidBody3DComponent>()) {
         const auto& tc = entity.Get<Transform3DComponent>();
         const auto& physics = entity.Get<RigidBody3DComponent>();
 
@@ -286,7 +326,7 @@ void PhysicsEngine3D::OnUpdate(Scene& scene, double dt)
     }
 
     // Post Update
-    for (auto entity : scene.Entities().View<RigidBody3DComponent>()) {
+    for (auto entity : registry.View<RigidBody3DComponent>()) {
         auto& tc = entity.Get<Transform3DComponent>();
         auto& rc = entity.Get<RigidBody3DComponent>();
         const rp3d::RigidBody* body = d_impl->entityData[entity].body;
